Fin debugage position

+++ Modele.cpp
@@ -19,7 +19,7 @@ double _size_cable_to_increment(double size_cable) {
 
 }
 
-void increment_moteur_from_pos(double pos_X_effecteur, double pos_Y_effecteur,double *increment_moteur_A, double *increment_moteur_B, double *increment_moteur_C) {
+void increment_moteur_from_pos(double pos_X_effecteur, double pos_Y_effecteur,int *increment_moteur_A, int *increment_moteur_B, int *increment_moteur_C) {
     double size_cable_A;
     double size_cable_B;
     double size_cable_C;
