+++ Modele.hpp
@@ -67,7 +67,7 @@ void _model_position_inverse(double pos_X_effecteur, double pos_Y_effecteur, dou
 double _size_cable_to_increment(double size_cable);
 
 
-void increment_moteur_from_pos(double pos_X_effecteur, double pos_Y_effecteur,double *increment_moteur_A, double *increment_moteur_B, double *increment_moteur_C);
+void increment_moteur_from_pos(double pos_X_effecteur, double pos_Y_effecteur,int *increment_moteur_A, int *increment_moteur_B, int *increment_moteur_C);
 
 //renvoie l'intersection entre deux cercles la plus proche du point C
 //tous les dimensions sont en mm
